+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.0.1
+
+### Bug Fixes
+
+- [#731](https://github.com/okta/okta-auth-js/pull/731) Fixes issue with `handleLoginRedirect` where a redirect could occur after an exception was thrown.
+  
 ## 5.0.0
 
 ### Features
@@ -20,6 +26,12 @@
 - [#675](https://github.com/okta/okta-auth-js/pull/675) Removes warning when calling `updateAuthState` when there are no subscribers
 - [#706](https://github.com/okta/okta-auth-js/pull/706) calling `isAuthenticated` will renew expired tokens when `autoRenew` is true
 
+## 4.9.1
+
+### Bug Fixes
+
+- [#731](https://github.com/okta/okta-auth-js/pull/731) Fixes issue with `handleLoginRedirect` where a redirect could occur after an exception was thrown.
+  
 ## 4.9.0
 
 ### Bug Fixes